requirements.txt
mysql-connector-python==8.0.29
paho-mqtt==1.6.1 